Add a small wrapper to create http requests
A small pet peeve of mine is mixing up data that changes at different
rates. For example, the device id isn't going to change very often,
whereas the list of contact events changes a lot while the application
is running. When possible I like to inject those at different times so
that variables that can be known ahead of time to the view controller
layer are hidden down below.

In this case I'm trying to hide the device id, since the PleaseIsolate
view controller doesn't really give a damn about that value at this
point in time, but the request needs it for constructing the appropriate
path that the server expects (e.g.: /api/residents/<uuid>).

Might be a little early to introduce such an abstraction but anything
that keeps the view controller layer clean will probably encourage us to
keep doing the right thing as we go along. I wouldn't be too upset if
this were deleted at some point.