Refactor : rename broadcast id generator to encrypter
I want to wrap up the necessary step of providing the sonarId and server
public key into a separate object that does the generation. That way the
encryption stays in one place and there is an easier to pass around
object to the bluetooth stack that can just call a method without
needing to have a bunch of other data on hand.