Switch broadcaster to use new broadcast format
Rather than a uuid, we now have the machinery in place to broadcast the
result of encrypting the uuid along with the current date against a
server id.

However, it's unfortunate that the server changes aren't ready yet. So
we have a hardcoded server key. Additionally, it requires flipping a
flag to use the new format, since I imagine it might not be very helpful
for the folks that are doing testing for me to switch this over now.