Clear out broadcast id encryption key from debug menu
This should be cleared at the same time as registration.
Makes me wonder whether this should be a part of the registration
object. Initially built it separate from that bit of code because
I was not sure how well this works, and I wanted a clean room to
experiment in from the rest of the app. I also wanted to ensure that
we could represent this data in two ways - when it is saved, as raw data,
and later on it should be returned as a SecKey. This might seem incidental
but I didn't want to litter the codebase with calls to SecKeyCreateWithData
and the like since those can fail, and it would be nice to consolidate
those in one place.

If you are reading this in the future and want to merge it into registration,
that's probably fine but be prepared for it to be painful.