Add a logger statement around guard in BLE broadcaster
I got really confused by this, so I'm writing a commit in order to help
jog my memory the next time I'm working on this bluetooth code - or
anyone else I suppose.

You might think that when the app starts up, we absolutely need to build
back up this state. Maybe you're debugging the bluetooth stack, and
you're trying to figure out why the new changes you want to be there
aren't present. This guard is important, do not delete it.

When debugging our app in xcode, if you terminate the app by stopping
the debugging session and it had previously been advertising, the
corebluetooth stack will restore the state of our service and it will
already be broadcasting by the time we startup. The best way to reset
the stack in that case is to toggle bluetooth off and on again, so that
we need to advertise our service again.