Use better bytes for coloring BLE debug view
Since the first byte is always 0x04, this would restrict the colors we
could render. Take bytes 2, 3, 4 instead.

Why is the first byte 0x04  you ask ? it's because for devices we
connect to, these bytes are actually the result of the encryption we do
to mask individual's sonar ids. As a result of the ECIES step, the
algorithm pads one byte at the beginning to indicate how the bytes for
the included ephemeral public key should be read. As far as I know this
is always 0x04 which indicates uncompressed point encoding. See the unit
tests for the id encrypter for more infomation.

Signed-off-by: Tim Jarratt <tjarratt@pivotal.io>