Ensure date based encryption tests pass
I had thought originally this was a timezone problem but in fact I was
forgetting to put this test in the right state. Since we support both
modes right now, this test would fail if we were using the old format,
since the uuid would be stable and never change.