Style backgrounds of onboarding screen
Removed the user consent toggle at the same time.

I was fairly unhappy with the cludgey code I had to write to get the
onboarding UI flow tests to pass. In short, the previous tests passed
because we persisted the user consent toggle value and forced the user
to go through that screen every single time.

Since the onboarding coordinator has no other way to distinguish the
"initial" state from the fact that the user may need to go to the
postcode screen, since we only ask it for its state, and it's allowed to
query state from persistence, there is no great way for it to know which
screen to present.

I decided to go with a throwaway boolean here in the onboarding
coordinator since that was the easiest thing that could work.

However, for the UI Flow tests, this needs to trigger some of the same
logic that was present in the onboarding view controller that the
RootViewController actually controls - ensuring we trigger the initial
state, by presenting the onboarding view controller inside the root vc.
If you remove that from the UITestScreenMaker, the tests fail because
you have to repeat the first two screens twice.

This is as much a pain point in the screen making ui test helper as it
is the onboarding coordinator. Those two could use some refactors to
make them a bit less fragile to this sort of bug.