Ensure we rotate our ephemeral broadcast id every day
Currently rotates at midnight, local time. However, it will also change
anytime the application gets restarted, or if the time changes
unexpectedly.

If you were to say change your phone's clock to be in the future, run
the app, change it back to the real time, and go back to the app, it
would not return to its original value, before you changed your phone's
clock to be set in the future. This is because we recompute the result,
and since it uses a new ephemeral keypair each time, we don't get the
same result.