Use beta firebase configuration
Unsure how this worked in earlier buidls but in the
preceding build it was impossible to register because
our backend services were receiving a 401 from firebase.
My understanding is that this is the result of our app's
bundle identifier differing from the configuration that was
present in firebase for the Internal builds.

I do not love calling this beta for now, I would much rather we
start calling it staging, to match the environment name, but for
now this will have to do

Signed-off-by: Anon 3 <anon3@example.com>