Do not focus registration success message for accessibility
This is problematic for accessibility reasons because in the successful
case, when registrations complete very quickly, there is an exceedingly
high chance that this message will be shown while a message is already
being read by VoiceOver. As a result, this impedes the WCAG guideline 4.1.3 AA
in that we would be impeding the user from what they were already doing.

For now, until we design a better version of this screen that doesn't
impede the user, we will not announce this change on the screen.